% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNUAL_INSTANT_PEAKS.R
\name{ANNUAL_INSTANT_PEAKS}
\alias{ANNUAL_INSTANT_PEAKS}
\title{Annual maximum/minimum instantaneous flows and water levels}
\usage{
ANNUAL_INSTANT_PEAKS(hydat_path, STATION_NUMBER = NULL,
  PROV_TERR_STATE_LOC = NULL, start_year = "ALL", end_year = "ALL")
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory.
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}

\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted from the function call, the value of \code{PROV_TERR_STATE_LOC}
is returned.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted from the function call, the value of \code{STATION_NUMBER}
is returned. See \code{unique(STATIONS(hydat_path = "H:/Hydat.sqlite3")$PROV_TERR_STATE_LOC)}}

\item{start_year}{First year of the returned record}

\item{end_year}{Last year of the returned record}
}
\value{
A tibble of ANNUAL_INSTANT_PEAKS
}
\description{
Provides wrapper to turn the ANNUAL_INSTANT_PEAKS table in HYDAT into a tidy data frame. \code{STATION_NUMBER} and
\code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{STATIONS} table are returned.
}
\examples{
\donttest{
## Multiple stations province not specified
ANNUAL_INSTANT_PEAKS(STATION_NUMBER = c("08NM083","08NE102"), hydat_path = "H:/Hydat.sqlite3")

## Multiple province, station number not specified
ANNUAL_INSTANT_PEAKS(PROV_TERR_STATE_LOC = c("AB","YT"), hydat_path = "H:/Hydat.sqlite3")
}

}
