% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNUAL_STATISTICS.R
\name{ANNUAL_STATISTICS}
\alias{ANNUAL_STATISTICS}
\title{Get a tidy tibble of daily flows}
\usage{
ANNUAL_STATISTICS(hydat_path = "H:/Hydat.sqlite3", STATION_NUMBER)
}
\arguments{
\item{hydat_path}{Directory to the hydat database}

\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument if all stations in BC are requested.}
}
\value{
A tibble of ANNUAL_STATISTICS
}
\description{
Provides wrapper to turn the ANNUAL_STATISTICS table into a tidy data frame
}
\examples{
ANNUAL_STATISTICS(STATION_NUMBER = "08LA001")

ANNUAL_STATISTICS(STATION_NUMBER = c("08LA001","08LG006"))
}
