% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_symbols}
\alias{data_symbols}
\title{DATA SYMBOLS look-up table}
\format{A tibble with 5 rows and 3 variables:
\describe{
  \item{SYMBOL_ID}{Symbol code}
  \item{SYMBOL_EN}{Description of Symbol (English)}
  \item{SYMBOL_FR}{Description of Symbol (French)}
}}
\source{
HYDAT
}
\usage{
data_symbols
}
\description{
A look table for data symbols
}
\keyword{datasets}
