% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_types}
\alias{data_types}
\title{DATA TYPES look-up table}
\format{A tibble with 5 rows and 3 variables:
\describe{
  \item{DATA_TYPE}{Data type code}
  \item{DATA_TYPE_EN}{Descriptive data type (English)}
  \item{DATA_TYPE_FR}{Descriptive data type (French)}
}}
\source{
HYDAT
}
\usage{
data_types
}
\description{
A look table for data types
}
\keyword{datasets}
