% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DATUM_LIST}
\alias{DATUM_LIST}
\title{DATUM_LIST function}
\usage{
DATUM_LIST(hydat_path)
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory}
}
\value{
A tibble of DATUMS
}
\description{
DATUM_LIST – DATUM look-up Table
}
