% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLY_LEVELS.R
\name{DLY_LEVELS}
\alias{DLY_LEVELS}
\title{Extract daily levels information from the HYDAT database}
\usage{
DLY_LEVELS(hydat_path, STATION_NUMBER, PROV_TERR_STATE_LOC,
  start_date = "ALL", end_date = "ALL")
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory}

\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. See also for argument options.}

\item{start_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}

\item{end_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}
}
\value{
A tibble of daily levels
}
\description{
Provides wrapper to turn the DLY_LEVELS table in HYDAT into a tidy data frame. \code{STATION_NUMBER} and 
\code{PROV_TERR_STATE_LOC} must both be supplied. When STATION_NUMBER="ALL" the PROV_TERR_STATE_LOC argument decides 
where those stations come from.
}
\examples{
\donttest{
DLY_LEVELS(STATION_NUMBER = "08LA001", PROV_TERR_STATE_LOC = "BC", 
           hydat_path = "H:/Hydat.sqlite3")
DLY_LEVELS(STATION_NUMBER = c("08LA001","08LG048"), PROV_TERR_STATE_LOC = "BC", 
          hydat_path = "H:/Hydat.sqlite3",
          start_date = "1996-01-01", end_date = "2000-01-01")

DLY_LEVELS(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")

DLY_LEVELS(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3",
          start_date = "1996-01-01", end_date = "2000-01-01")
          }

}
\seealso{
Possible arguments for \code{PROV_TERR_STATE_LOC}
\itemize{
\item "QC" 
\item "ME" 
\item "NB" 
\item "PE" 
\item "NS" 
\item "MN" 
\item "ON" 
\item "MI" 
\item "NL" 
\item "MB" 
\item "AB" 
\item "MT" 
\item "SK" 
\item "ND" 
\item "NU" 
\item "NT" 
\item "BC" 
\item "YT" 
\item "AK" 
\item "WA" 
\item "ID"
}
}
