% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_realtime}
\alias{download_realtime}
\title{Download a tibble of realtime network}
\usage{
download_realtime(STATION_NUMBER, PROV_TERR_STATE_LOC = "BC")
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number. Multiple stations can inputted. See examples}

\item{PROV_TERR_STATE_LOC}{Jurisdiction where the station is located. Functionality is currently limited. Defaults to BC. 
Other jurisdictions can be used but the STATION_NUMBER argument must be located in that jurisdiction.}
}
\value{
A tibble of water flow and level values
}
\description{
A function to download realtime discharge data from the Water Survey of Canada datamart. Multiple stations will
be used. Currently, if a station does not exist or is not found, no data is returned.
}
\note{
This function is heavily adapted from the RealTimeData function from the HYDAT package. 
That package can be viewed here: \url{https://github.com/CentreForHydrology/HYDAT}.
Differences between HYDAT::RealTimeData and download_realtime include
\itemize{
\item Column name outputted by download_realtime are identical to corresponding measures in HYDAT
\item Uses readr::read_csv and outputs a tibble
\item download_network is approximately 50 times faster than HYDAT::RealTimeNetwork
}
}
\examples{
download_realtime(STATION_NUMBER="08MF005")

download_realtime(STATION_NUMBER=c("08MF005", "08NL071"))

}
