% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_realtime2}
\alias{download_realtime2}
\title{Download a tibble of realtime discharge data from the MSC datamart}
\usage{
download_realtime2(STATION_NUMBER, PROV_TERR_STATE_LOC)
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. See also for argument options.}
}
\value{
A tibble of water flow and level values
}
\description{
Download realtime discharge data from the Meteorological Service of Canada (MSC) datamart. The function will prioritize 
downloading data collected at the highest resolution. In instances where data is not available at high (hourly or higher) resolution 
daily averages are used. Currently, if a station does not exist or is not found, no data is returned. Both the province and the station number 
should be specified.
}
\examples{
download_realtime2(STATION_NUMBER="08MF005", PROV_TERR_STATE_LOC="BC")

# To download all stations in Prince Edward Island:
download_realtime2(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE")

}
\seealso{
\code{download_network()}.
}
