% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_realtime_ws}
\alias{download_realtime_ws}
\title{Download realtime data from the ECCC web service
Function to actually retrieve data from ECCC webservice. Before using this function,
a token from \code{get_ws_token()} is needed.}
\usage{
download_realtime_ws(STATION_NUMBER, parameters = c(46, 16, 52, 47, 8, 5, 41,
  18), start_date = Sys.Date() - 30, end_date = Sys.Date(), token)
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number.}

\item{parameters}{parameter ID. Can take multiple entries. Parameter is a numeric code. See \code{param_id} for options. Defaults to all parameters.}

\item{start_date}{Need to be in YYYY-MM-DD. Defaults to 30 days before current date}

\item{end_date}{Need to be in YYYY-MM-DD. Defaults to current date}

\item{token}{generate by \code{get_ws_token()}}
}
\description{
Download realtime data from the ECCC web service
Function to actually retrieve data from ECCC webservice. Before using this function,
a token from \code{get_ws_token()} is needed.
}
\examples{
\donttest{
token_out <- get_ws_token(username = Sys.getenv("WS_USRNM"), password = Sys.getenv("WS_PWD"))

ws_08 <- download_realtime_ws(STATION_NUMBER = c("08NL071","08NM174"),
                         parameters = c(47, 5),
                         token = token_out)

fivedays <- download_realtime_ws(STATION_NUMBER = c("08NL071","08NM174"),
                         parameters = c(47, 5),
                         end_date = Sys.Date(), #today
                         start_date = Sys.Date() - 5, #five days ago
                         token = token_out)
}
}
\seealso{
Other realtime functions: \code{\link{download_realtime_dd}},
  \code{\link{get_ws_token}},
  \code{\link{realtime_network_meta}}
}
