% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ws}
\alias{download_ws}
\title{Download data from the ECCC web service}
\usage{
download_ws(STATION_NUMBER, parameters, start_date, end_date, token)
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number.}

\item{parameters}{parameter ID}

\item{start_date}{Need to be in YYYY-MM-DD}

\item{end_date}{Need to be in YYYY-MM-DD}

\item{token}{generate by \code{get_ws_token()}}
}
\description{
Function to actually retrieve data from ECCC webservice. Before using this function, 
a token from \code{get_ws_token()} is needed.
}
