% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_annual_instant_peaks.R
\name{hy_annual_instant_peaks}
\alias{hy_annual_instant_peaks}
\title{Extract annual max/min instantaneous flows and water levels from HYDAT database}
\source{
HYDAT
}
\usage{
hy_annual_instant_peaks(
  station_number = NULL,
  hydat_path = NULL,
  prov_terr_state_loc = NULL,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{station_number}{A seven digit Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{hydat_path}{The path to the hydat database or NULL to use the default location
used by \link{download_hydat}. It is also possible to pass in an existing
database connection from \code{\link[=hy_src]{hy_src()}} such that the database only needs to be opened once per
user-level call.}

\item{prov_terr_state_loc}{Province, state or territory. If this argument is omitted, the value of \code{station_number}
is returned. See \code{unique(allstations$prov_terr_state_loc)}. Will also accept \code{CA} to return only Canadian stations.}

\item{start_year}{First year of the returned record}

\item{end_year}{Last year of the returned record}
}
\value{
A tibble of hy_annual_instant_peaks.
}
\description{
Provides wrapper to turn the ANNUAL_INSTANT_PEAKS table in HYDAT into a tidy data frame of instantaneous flows and water levels.
\code{station_number} and \code{prov_terr_state_loc} can both be supplied.
}
\examples{
\dontrun{
## Multiple stations province not specified
hy_annual_instant_peaks(station_number = c("08NM083", "08NE102"))

## Multiple province, station number not specified
hy_annual_instant_peaks(prov_terr_state_loc = c("AB", "YT"))
}

}
\seealso{
Other HYDAT functions:
\code{\link[=hy_agency_list]{hy_agency_list()}},
\code{\link[=hy_annual_stats]{hy_annual_stats()}},
\code{\link[=hy_daily]{hy_daily()}},
\code{\link[=hy_daily_flows]{hy_daily_flows()}},
\code{\link[=hy_daily_levels]{hy_daily_levels()}},
\code{\link{hy_data_symbols}},
\code{\link{hy_data_types}},
\code{\link[=hy_datum_list]{hy_datum_list()}},
\code{\link[=hy_monthly_flows]{hy_monthly_flows()}},
\code{\link[=hy_monthly_levels]{hy_monthly_levels()}},
\code{\link[=hy_reg_office_list]{hy_reg_office_list()}},
\code{\link[=hy_sed_daily_loads]{hy_sed_daily_loads()}},
\code{\link[=hy_sed_daily_suscon]{hy_sed_daily_suscon()}},
\code{\link[=hy_sed_monthly_loads]{hy_sed_monthly_loads()}},
\code{\link[=hy_sed_monthly_suscon]{hy_sed_monthly_suscon()}},
\code{\link[=hy_sed_samples]{hy_sed_samples()}},
\code{\link[=hy_sed_samples_psd]{hy_sed_samples_psd()}},
\code{\link[=hy_stations]{hy_stations()}},
\code{\link[=hy_stn_data_coll]{hy_stn_data_coll()}},
\code{\link[=hy_stn_data_range]{hy_stn_data_range()}},
\code{\link[=hy_stn_op_schedule]{hy_stn_op_schedule()}},
\code{\link[=hy_stn_regulation]{hy_stn_regulation()}},
\code{\link[=hy_version]{hy_version()}}
}
\concept{HYDAT functions}
