% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_stn_.R
\name{hy_stn_datum_conv}
\alias{hy_stn_datum_conv}
\title{hy_stn_datum_conv function}
\format{A tibble with 4 variables:
\describe{
  \item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
  \item{DATUM_EN_FROM}{Identifying a datum from which water level is being converted}
  \item{DATUM_EN_TO}{Identifying a datum to which water level is being converted}
  \item{CONVERSTION_FACTOR}{The conversion factor applied to water levels referred to 
  one datum to obtain water levels referred to another datum}
}}
\usage{
hy_stn_datum_conv(station_number = NULL, hydat_path = NULL,
  prov_terr_state_loc = NULL)
}
\arguments{
\item{station_number}{A seven digits Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{hydat_path}{The default for this argument is to look for hydat in the same location where it
was saved by using \code{download_hydat}. Therefore this argument is almost always omitted from a function call. 
You can see where hydat was downloaded using \code{hy_dir()}}

\item{prov_terr_state_loc}{Province, state or territory. If this argument is omitted, the value of \code{station_number}
is returned. See \code{unique(allstations$prov_terr_state_loc)}}
}
\value{
A tibble of hy_stn_datum_conv
}
\description{
hy_stn_datum_conv look-up Table
}
\examples{
\dontrun{
hy_stn_datum_conv(station_number = c("02JE013","08MF005"))
}
}
