% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STN_.R
\name{hy_stn_remarks}
\alias{hy_stn_remarks}
\title{hy_stn_remarks function}
\usage{
hy_stn_remarks(STATION_NUMBER = NULL,
  hydat_path = paste0(rappdirs::user_data_dir(), "\\\\Hydat.sqlite3"),
  PROV_TERR_STATE_LOC = NULL)
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted, the value of \code{PROV_TERR_STATE_LOC}
is returned.}

\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory.
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted, the value of \code{STATION_NUMBER}
is returned. See \code{unique(allstations$PROV_TERR_STATE_LOC)}}
}
\value{
A tibble of hy_stn_remarks
}
\description{
hy_stn_remarks look-up Table
}
\examples{
\donttest{
hy_stn_remarks(STATION_NUMBER = c("02JE013","08MF005"))
}

}
