% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_db.R
\name{hy_test_db}
\alias{hy_test_db}
\title{Get the location of the test HYDAT database}
\usage{
hy_test_db()
}
\value{
The file location of tinyhydat.sqlite3
}
\description{
The full HYDAT database needs to be downloaded from \link{download_hydat}, but for testing
purposes, a small test database is included in this package. Use
\code{hydat_path = hy_test_db()} in hy_* functions to use the test database.
}
\examples{
hy_test_db()

}
\seealso{
\link{hy_db}
}
