% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_id}
\alias{param_id}
\title{Parameter ID}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 7 columns.}
\source{
\url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometric_additionalData/Document/WebService_Guidelines.pdf}
}
\usage{
param_id
}
\description{
A tibble of parameter id codes and their corresponding explanation/description
}
\keyword{datasets}
