% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_plot.R, R/realtime_plot.R
\name{plot}
\alias{plot}
\alias{plot.hy}
\alias{plot.realtime}
\title{plot hy object}
\usage{
\method{plot}{hy}(x = NULL, ...)

\method{plot}{realtime}(x = NULL, Parameter = c("Flow", "Level"), ...)
}
\arguments{
\item{x}{Object created by either a hy_daily_* or realtime_dd data retrieval function}

\item{...}{passed to \code{plot}}

\item{Parameter}{Parameter of interest. Either "Flow" or "Level". Defaults to "Flow".}
}
\description{
This method plots daily time series data from HYDAT. Functionality is very basic
and is intended for exploratory purposes only.

This method plots daily time series data from the ECCC datamart. Functionality is very basic
and is intended for exploratory purposes only.
}
\examples{
\dontrun{
# One station
fraser <- hy_daily_flows("08MF005")
plot(fraser)
}

\dontrun{
# One station
fraser <- realtime_dd("08MF005")
plot(fraser)
}

}
