% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{realtime_daily_mean}
\alias{realtime_daily_mean}
\title{Calculate daily means from higher resolution realtime data}
\usage{
realtime_daily_mean(.data, na.rm = FALSE)
}
\arguments{
\item{.data}{A data argument that is designed to take only the output of realtime_dd}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
This function is meant to be used within a pipe as a means of easily moving from higher resolution
data to daily means.
}
\examples{
\dontrun{
realtime_dd("08MF005") \%>\% realtime_daily_mean()
}

}
