% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{realtime_dd}
\alias{realtime_dd}
\title{Download a tibble of realtime river data from the last 30 days from the Meteorological Service of Canada datamart}
\format{
A tibble with 8 variables:
\itemize{
\item \strong{STATION_NUMBER}: Unique 7 digit Water Survey of Canada station number
\item \strong{PROV_TERR_STATE_LOC}: The province, territory or state in which the station is located
\item \strong{Date}: Observation date and time for last thirty days. Formatted as a POSIXct class in UTC for consistency.
\item \strong{Parameter}: Parameter being measured. Only possible values are Flow and Level
\item \strong{Value}: Value of the measurement. If Parameter equals Flow the units are m^3/s.
If Parameter equals Level the units are metres.
\item \strong{Grade}: reserved for future use
\item \strong{Symbol}: reserved for future use
\item \strong{Code}: quality assurance/quality control flag for the discharge
\item \strong{station_tz}: Station timezone based on tidyhydat::allstations$station_tz
}
}
\usage{
realtime_dd(station_number = NULL, prov_terr_state_loc = NULL)
}
\arguments{
\item{station_number}{A seven digit Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{prov_terr_state_loc}{Province, state or territory. If this argument is omitted, the value of \code{station_number}
is returned. See \code{unique(allstations$prov_terr_state_loc)}. Will also accept \code{CA} to return only Canadian stations.}
}
\value{
A tibble of water flow and level values. The date and time of the query (in UTC) is also
stored as an attribute.
}
\description{
Download realtime river data from the last 30 days from the Meteorological Service of Canada (MSC) datamart.
The function will prioritize downloading data collected at the highest resolution. In instances where data is
not available at high (hourly or higher) resolution daily averages are used. Currently, if a station does not
exist or is not found, no data is returned.
}
\examples{
\dontrun{
## Download from multiple provinces
realtime_dd(station_number = c("01CD005", "08MF005"))

## To download all stations in Prince Edward Island:
pei <- realtime_dd(prov_terr_state_loc = "PE")

## Access the time of query
attributes(pei)$query_time
}

}
\seealso{
Other realtime functions:
\code{\link[=realtime_stations]{realtime_stations()}},
\code{\link[=realtime_ws]{realtime_ws()}}
}
\concept{realtime functions}
