% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{REGIONAL_OFFICE_LIST}
\alias{REGIONAL_OFFICE_LIST}
\title{REGIONAL_OFFICE_LIST function}
\usage{
REGIONAL_OFFICE_LIST(hydat_path)
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory}
}
\value{
A tibble of offices
}
\description{
REGIONAL_OFFICE_LIST – OFFICE look-up Table
}
