% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SED_MONTHLY_LOADS.R
\name{SED_MONTHLY_LOADS}
\alias{SED_MONTHLY_LOADS}
\title{Extract monthly flows information from the HYDAT database}
\source{
HYDAT
}
\usage{
SED_MONTHLY_LOADS(hydat_path = NULL, STATION_NUMBER = NULL,
  PROV_TERR_STATE_LOC = NULL, start_date = "ALL", end_date = "ALL")
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory.
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}

\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted from the function call, the value of \code{PROV_TERR_STATE_LOC}
is returned.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted from the function call, the value of \code{STATION_NUMBER}
is returned. See \code{unique(STATIONS(hydat_path = "H:/Hydat.sqlite3")$PROV_TERR_STATE_LOC)}}

\item{start_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}

\item{end_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}
}
\value{
A tibble of monthly flows. This includes a \code{Date_occured} column which indicates the date of the \code{Sum_stat}. For MEAN and 
  TOTAL this is not presented as those aren't daily values.
}
\description{
Provides wrapper to turn the SED_MONTHLY_LOADS table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
  \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{STATIONS} table are returned.
  That is a large vector for \code{SED_MONTHLY_LOADS}.
}
\examples{
\donttest{
SED_MONTHLY_LOADS(STATION_NUMBER = c("02JE013","08MF005"), hydat_path = "H:/Hydat.sqlite3",
start_date = "1996-01-01", end_date = "2000-01-01")

SED_MONTHLY_LOADS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
          }
}
\seealso{
Other HYDAT functions: \code{\link{MONTHLY_FLOWS}},
  \code{\link{MONTHLY_LEVELS}},
  \code{\link{SED_MONTHLY_SUSCON}}
}
