% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{station_choice}
\alias{station_choice}
\title{Function to chose a station based on consistent arguments for hydat functions.}
\usage{
station_choice(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC)
}
\arguments{
\item{hydat_con}{A database connection}

\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted, the value of \code{PROV_TERR_STATE_LOC}
is returned.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted, the value of \code{STATION_NUMBER}
is returned. See \code{unique(allstations$PROV_TERR_STATE_LOC)}}
}
\description{
A function to avoid duplication in HYDAT functions.  This function is not intended for external use.
}
\keyword{internal}
