% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STN_.R
\name{STN_DATA_COLLECTION}
\alias{STN_DATA_COLLECTION}
\title{STN_DATA_COLLECTION function}
\source{
HYDAT
}
\usage{
STN_DATA_COLLECTION(hydat_path = NULL, STATION_NUMBER = NULL,
  PROV_TERR_STATE_LOC = NULL)
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory.
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}

\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted, the value of \code{PROV_TERR_STATE_LOC}
is returned.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted, the value of \code{STATION_NUMBER}
is returned. See \code{unique(allstations$PROV_TERR_STATE_LOC)}}
}
\value{
A tibble of STN_DATA_COLLECTION
}
\description{
STN_DATA_COLLECTION look-up Table
}
\examples{
\donttest{
STN_DATA_COLLECTION(STATION_NUMBER = c("02JE013","08MF005"))
}

}
\seealso{
Other HYDAT functions: \code{\link{AGENCY_LIST}},
  \code{\link{ANNUAL_INSTANT_PEAKS}},
  \code{\link{ANNUAL_STATISTICS}},
  \code{\link{DATUM_LIST}}, \code{\link{DLY_FLOWS}},
  \code{\link{DLY_LEVELS}}, \code{\link{MONTHLY_FLOWS}},
  \code{\link{MONTHLY_LEVELS}},
  \code{\link{REGIONAL_OFFICE_LIST}},
  \code{\link{SED_DLY_LOADS}},
  \code{\link{SED_DLY_SUSCON}},
  \code{\link{SED_MONTHLY_LOADS}},
  \code{\link{SED_MONTHLY_SUSCON}},
  \code{\link{SED_SAMPLES_PSD}}, \code{\link{SED_SAMPLES}},
  \code{\link{STATIONS}}, \code{\link{STN_DATA_RANGE}},
  \code{\link{STN_OPERATION_SCHEDULE}},
  \code{\link{STN_REGULATION}}, \code{\link{VERSION}}
}
