% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{token_ws}
\alias{token_ws}
\title{Request a token from the Environment and Climate Change Canada [EXPERIMENTAL]}
\usage{
token_ws(username = NULL, password = NULL)
}
\arguments{
\item{username}{Supplied by ECCC. Defaults to NULL which will use WS_USRNM from .Renviron file}

\item{password}{Supplied by ECCC. Defaults to NULL which will use WS_PWD from .Renviron file
Request a token from the ECCC web service using the POST method. This token expires after 10 minutes.
You can only have 5 tokens out at once.}
}
\value{
The token as a string that should be supplied the \code{download_ws_realtime} function.
}
\description{
Request a token from the Environment and Climate Change Canada [EXPERIMENTAL]
}
\details{
The \code{username} and \code{password} should be treated carefully and should never be entered directly into an r script or console.
Rather these credentials should be stored in your \code{.Renviron} file. The .Renviron file can edited using \code{file.edit("~/.Renviron")}.
In that file, which is only stored locally and is only available to you, you can assign your \code{username} to WS_USRNM and \code{password} 
to WS_PWD and then simply issue \code{token_ws}. See \code{?download_ws} for examples.
}
\seealso{
Other realtime functions: \code{\link{realtime_dd}},
  \code{\link{realtime_stations}},
  \code{\link{realtime_ws}}
}
