% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical-webservice.R
\name{ws_daily_flows}
\alias{ws_daily_flows}
\alias{ws_daily_levels}
\title{Download historical flow and level data from the ECCC web service}
\format{
A tibble with 6 variables:
\describe{
\item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
\item{Date}{Observation date and time. Formatted as a POSIXct class as UTC for consistency.}
\item{Parameter}{Type of parameter}
\item{Value}{Value of the measurement.}
\item{Symbol}{future use}
}
}
\usage{
ws_daily_flows(station_number, start_date = NULL, end_date = NULL)

ws_daily_levels(station_number, start_date = NULL, end_date = NULL)
}
\arguments{
\item{station_number}{Water Survey of Canada station number.}

\item{start_date}{Accepts YYYY-MM-DD. You need to provide a start date.
The default value is NULL}

\item{end_date}{Accepts either YYYY-MM-DD. You need to provide an end date.
The default value is NULL}
}
\description{
Functions to retrieve historical flow and levels data from ECCC web service. This data is
the same as HYDAT data but provides the convenience of not having to download
the HYDAT database. This function is useful when a smaller amount of data is needed. If
you need lots of data, consider using HYDAT and the \code{hy_} family of functions
}
\examples{
\dontrun{
try(
   flow_data <- ws_daily_flows(
     station_number = c("08NL071", "08NM174"),
     start_date = Sys.Date() - 365,
     end_date = Sys.Date()
   )
)
try(
   level_data <- ws_daily_level(
     station_number = c("08NL071", "08NM174"),
     start_date = Sys.Date() - 365,
     end_date = Sys.Date()
     )
)
}
}
\seealso{
hy_daily_flows
}
