% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allstations}
\alias{allstations}
\title{A tibble of all Canadian Stations stations and their names.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7791 rows and 3 columns.}
\source{
HYDAT
}
\usage{
allstations
}
\description{
A shorthand to avoid having always call \code{hy_stations}. Only up to date as of 2017-07-18.
}
\keyword{datasets}
