% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNUAL_STATISTICS.R
\name{ANNUAL_STATISTICS}
\alias{ANNUAL_STATISTICS}
\title{Extract daily flows information from the HYDAT database}
\usage{
ANNUAL_STATISTICS(hydat_path = "H:/Hydat.sqlite3", STATION_NUMBER,
  PROV_TERR_STATE_LOC, start_year = "ALL", end_year = "ALL")
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory}

\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. See also for argument options.}

\item{start_year}{First year of the returned record}

\item{end_year}{Last year of the returned record}
}
\value{
A tibble of ANNUAL_STATISTICS
}
\description{
Provides wrapper to turn the ANNUAL_STATISTICS table in HYDAT into a tidy data frame. \code{STATION_NUMBER} and 
\code{PROV_TERR_STATE_LOC} must both be supplied. When STATION_NUMBER="ALL" the PROV_TERR_STATE_LOC argument decides 
where those stations come from.
}
\examples{
\donttest{
ANNUAL_STATISTICS(STATION_NUMBER = "08LA001",
                  PROV_TERR_STATE_LOC = "BC", hydat_path = "H:/Hydat.sqlite3")

ANNUAL_STATISTICS(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE", 
                  hydat_path = "H:/Hydat.sqlite3")

ANNUAL_STATISTICS(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE", 
                  hydat_path = "H:/Hydat.sqlite3",
                  start_year = 1972,
                  end_year = 1975)
}
                  
}
\seealso{
Possible arguments for \code{PROV_TERR_STATE_LOC}
\itemize{
\item "QC" 
\item "ME" 
\item "NB" 
\item "PE" 
\item "NS" 
\item "MN" 
\item "ON" 
\item "MI" 
\item "NL" 
\item "MB" 
\item "AB" 
\item "MT" 
\item "SK" 
\item "ND" 
\item "NU" 
\item "NT" 
\item "BC" 
\item "YT" 
\item "AK" 
\item "WA" 
\item "ID"
}
}
