% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATA_SYMBOLS}
\alias{DATA_SYMBOLS}
\title{DATA_SYMBOLS – DATA SYMBOLS look-up table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.}
\source{
HYDAT
}
\usage{
DATA_SYMBOLS
}
\description{
A look table for data symbols
}
\keyword{datasets}
