% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_types}
\alias{data_types}
\title{DATA TYPES look-up table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.}
\source{
HYDAT
}
\usage{
data_types
}
\description{
A look table for data types
}
\keyword{datasets}
