% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLY_FLOWS.R
\name{DLY_FLOWS}
\alias{DLY_FLOWS}
\title{Get a tidy tibble of daily flows}
\usage{
DLY_FLOWS(hydat_path = "H:/Hydat.sqlite3", STATION_NUMBER,
  PROV_TERR_STATE_LOC)
}
\arguments{
\item{hydat_path}{Directory to the hydat database}

\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument at which point you need 
to specify \code{PROV_TERR_STATE_LOC}.}

\item{PROV_TERR_STATE_LOC}{Can be any province. See also for argument options.}
}
\value{
A tibble of daily flows
}
\description{
Provides wrapper to turn the DLY_FLOWS table into a tidy tibble
}
\examples{
DLY_FLOWS(STATION_NUMBER = "08LA001", PROV_TERR_STATE_LOC = "BC")

DLY_FLOWS(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE")

}
\seealso{
Possible arguments for \code{PROV_TERR_STATE_LOC}
\itemize{
\item "QC" 
\item "ME" 
\item "NB" 
\item "PE" 
\item "NS" 
\item "MN" 
\item "ON" 
\item "MI" 
\item "NL" 
\item "MB" 
\item "AB" 
\item "MT" 
\item "SK" 
\item "ND" 
\item "NU" 
\item "NT" 
\item "BC" 
\item "YT" 
\item "AK" 
\item "WA" 
\item "ID"
}
}
