% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLY_FLOWS.R
\name{DLY_FLOWS}
\alias{DLY_FLOWS}
\title{Extract daily flows information from the HYDAT database}
\source{
HYDAT
}
\usage{
DLY_FLOWS(hydat_path = NULL, STATION_NUMBER = NULL,
  PROV_TERR_STATE_LOC = NULL, start_date = "ALL", end_date = "ALL")
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory.
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}

\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted, the value of \code{PROV_TERR_STATE_LOC}
is returned.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted, the value of \code{STATION_NUMBER}
is returned. See \code{unique(allstations$PROV_TERR_STATE_LOC)}}

\item{start_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}

\item{end_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}
}
\value{
A tibble of daily levels
}
\description{
Provides wrapper to turn the DLY_FLOWS table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
\code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{STATIONS} table are returned.
That is a large vector for \code{DLY_FLOWS}.
}
\examples{
\donttest{
DLY_FLOWS(STATION_NUMBER = c("02JE013","08MF005"), hydat_path = "H:/Hydat.sqlite3",
start_date = "1996-01-01", end_date = "2000-01-01")

DLY_FLOWS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")

          }

}
\seealso{
Other HYDAT functions: \code{\link{AGENCY_LIST}},
  \code{\link{ANNUAL_INSTANT_PEAKS}},
  \code{\link{ANNUAL_STATISTICS}},
  \code{\link{DATUM_LIST}}, \code{\link{DLY_LEVELS}},
  \code{\link{MONTHLY_FLOWS}},
  \code{\link{MONTHLY_LEVELS}},
  \code{\link{REGIONAL_OFFICE_LIST}},
  \code{\link{SED_DLY_LOADS}},
  \code{\link{SED_DLY_SUSCON}},
  \code{\link{SED_MONTHLY_LOADS}},
  \code{\link{SED_MONTHLY_SUSCON}},
  \code{\link{SED_SAMPLES_PSD}}, \code{\link{SED_SAMPLES}},
  \code{\link{STATIONS}},
  \code{\link{STN_DATA_COLLECTION}},
  \code{\link{STN_DATA_RANGE}},
  \code{\link{STN_OPERATION_SCHEDULE}},
  \code{\link{STN_REGULATION}}, \code{\link{VERSION}}
}
