% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLY_FLOWS.R
\name{DLY_FLOWS}
\alias{DLY_FLOWS}
\title{Get a tidy tibble of daily flows}
\usage{
DLY_FLOWS(hydat_path = "H:/Hydat.sqlite3",
  STATION_NUMBER_SEL = STATION_NUMBER)
}
\arguments{
\item{hydat_path}{Directory to the hydat database}

\item{STATION_NUMBER_SEL}{Water Survey of Canada station number}
}
\description{
Provides wrapper to turn the DLY_FLOWS table into a tidy data frame
}
\examples{
DLY_FLOWS(STATION_NUMBER = "08LA001")

DLY_FLOWS(STATION_NUMBER = c("08LA001","08LG006"))
}
