% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLY_LEVELS.R
\name{DLY_LEVELS}
\alias{DLY_LEVELS}
\title{Extract daily levels information from the HYDAT database}
\usage{
DLY_LEVELS(hydat_path, STATION_NUMBER = NULL, PROV_TERR_STATE_LOC = NULL,
  start_date = "ALL", end_date = "ALL")
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory}

\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted from the function call, the value of \code{PROV_TERR_STATE_LOC} 
is returned.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted from the function call, the value of \code{STATION_NUMBER} 
is returned. See \code{unique(STATIONS(hydat_path = "H:/Hydat.sqlite3")$PROV_TERR_STATE_LOC)}}

\item{start_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}

\item{end_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}
}
\value{
A tibble of daily levels
}
\description{
Provides wrapper to turn the DLY_LEVELS table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and 
\code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{STATIONS} table are returned.
That is a large vector for \code{DLY_LEVELS}.
}
\examples{
\donttest{
DLY_LEVELS(STATION_NUMBER = c("02JE013","08MF005"), hydat_path = "H:/Hydat.sqlite3",
start_date = "1996-01-01", end_date = "2000-01-01")

DLY_LEVELS(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")

          }

}
