% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_hydat}
\alias{download_hydat}
\title{Download and set the path to HYDAT}
\usage{
download_hydat(dl_hydat_here = NULL)
}
\arguments{
\item{dl_hydat_here}{Directory to the HYDAT database. The path is chosen by the \code{rappdirs} package and is OS specific and can be view by \code{hy_dir}. 
This path is also supplied automatically to any function that uses the HYDAT database. A user specified path can be set though this is not the advised approach. 
It also downloads the database to a directory specified by \code{hy_dir}.}
}
\description{
Download the HYDAT sqlite database. This database contains all the historical hydrometric data for Canada's integrated hydrometric network.
The function will check for a existing sqlite file and won't download the file if the same version is already present.
}
\examples{
\dontrun{
download_hydat()
}

}
