% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_realtime2}
\alias{download_realtime2}
\title{Download a tibble of realtime network discharge data from the datamart}
\usage{
download_realtime2(STATION_NUMBER, PROV_TERR_STATE_LOC)
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. See also for argument options.}
}
\value{
A tibble of water flow and level values
}
\description{
A function to download realtime discharge data from the Water Survey of Canada datamart. Multiple stations will
be used. Currently, if a station does not exist or is not found, no data is returned. Both the province and the station number 
should be specified.
}
\note{
This function is inspired from the RealTimeData function from the HYDAT package. 
That package can be viewed here: \url{https://github.com/CentreForHydrology/HYDAT}.
Differences between HYDAT::RealTimeData and download_realtime include
\itemize{
\item Column name outputted by download_realtime are identical to corresponding measures in HYDAT
\item Uses readr::read_csv and outputs a tibble
\item download_network is approximately 50% times faster than HYDAT::RealTimeNetwork
}
}
\examples{
download_realtime2(STATION_NUMBER="08MF005", PROV_TERR_STATE_LOC="BC")

# To download all stations in Prince Edward Island:
download_realtime2(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE")

}
\seealso{
\code{download_network()}.
}
