% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_ws_token}
\alias{get_ws_token}
\title{Request a token from the Environment and Climate Change Canada}
\usage{
get_ws_token(username, password)
}
\arguments{
\item{username}{Supplied by ECCC}

\item{password}{Supplied by ECCC}
}
\value{
The token as a string that should be supplied the \code{download_ws_realtime} function.
}
\description{
Request a token from the ECCC webservice using the POST method. This token expires after 10 minutes. 
You can only have 5 tokens out at once.
}
