% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy.R
\name{hy_dir}
\alias{hy_dir}
\title{Output OS-independent path to the HYDAT sqlite database}
\usage{
hy_dir(...)
}
\arguments{
\item{...}{arguments potentially passed to \code{rappdirs::user_data_dir}}
}
\description{
Provides the download location for \link{download_hydat} in an OS independent manner.
}
\examples{
\dontrun{
hy_dir()
}

}
