% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sed_monthly_suscon.R
\name{hy_sed_monthly_suscon}
\alias{hy_sed_monthly_suscon}
\title{Extract monthly flows information from the HYDAT database}
\source{
HYDAT
}
\usage{
hy_sed_monthly_suscon(station_number = NULL,
  hydat_path = paste0(rappdirs::user_data_dir(), "\\\\Hydat.sqlite3"),
  prov_terr_state_loc = NULL, start_date = "ALL", end_date = "ALL")
}
\arguments{
\item{station_number}{Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory.
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}

\item{prov_terr_state_loc}{Province, state or territory. If this argument is omitted, the value of \code{station_number}
is returned. See \code{unique(allstations$prov_terr_state_loc)}}

\item{start_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}

\item{end_date}{Leave blank if all dates are required. Date format needs to be in YYYY-MM-DD. Date is inclusive.}
}
\value{
A tibble of monthly suspended sediment concentrations. This includes a \code{Date_occured} column which indicates the date of the \code{Sum_stat}. For MEAN and 
  TOTAL this is not presented as those are not daily values.
}
\description{
Provides wrapper to turn the hy_sed_monthly_suscon table in HYDAT into a tidy data frame.  \code{station_number} and
  \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
  That is a large vector for \code{hy_sed_monthly_suscon}.
}
\examples{
\donttest{
hy_sed_monthly_suscon(station_number = "08MF005")
          }
}
\seealso{
Other HYDAT functions: \code{\link{hy_agency_list}},
  \code{\link{hy_annual_instant_peaks}},
  \code{\link{hy_annual_stats}},
  \code{\link{hy_daily_flows}},
  \code{\link{hy_daily_levels}},
  \code{\link{hy_datum_list}},
  \code{\link{hy_monthly_flows}},
  \code{\link{hy_monthly_levels}},
  \code{\link{hy_reg_office_list}},
  \code{\link{hy_sed_daily_loads}},
  \code{\link{hy_sed_daily_suscon}},
  \code{\link{hy_sed_monthly_loads}},
  \code{\link{hy_sed_samples_psd}},
  \code{\link{hy_sed_samples}}, \code{\link{hy_stations}},
  \code{\link{hy_stn_data_coll}},
  \code{\link{hy_stn_data_range}},
  \code{\link{hy_stn_op_schedule}},
  \code{\link{hy_stn_regulation}}, \code{\link{hy_version}}
}
