% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_db.R
\name{hy_src}
\alias{hy_src}
\alias{hy_db}
\title{Open a connection to the HYDAT database}
\usage{
hy_src(hydat_path = NULL)

hy_db(hydat_path = NULL, check_exists = TRUE)
}
\arguments{
\item{hydat_path}{The path to the hydat database or NULL to use the default location
used by \link{download_hydat}.}

\item{check_exists}{Throw an error if hydat_path or the default database do not exist.}
}
\value{
\code{hy_src} returns a dplyr \link[dplyr]{src_sqlite}; 
  \code{hy_db} returns the file locations of the downloaded HYDAT database
}
\description{
This function gives low-level access to the underlying HYDAT database used by
other functions. Many of these tables are too large to load into memory,
so it is best to use dplyr to \link[dplyr]{filter} them before using
\link[dplyr]{collect} to read them into memory.
}
\examples{
library(dplyr)

# src is a src_sqlite
src <- hy_src(hydat_path = hy_test_db())
src_tbls(src)

# to get a table, use dplyr::tbl()
tbl(src, "STATIONS")

# one you're sure the results are what you want
# get a data.frame using collect()
tbl(src, "STATIONS") \%>\%
  filter(PROV_TERR_STATE_LOC == "BC") \%>\%
  collect()
  
# close the connection to the database by removing the object
# (and triggering garbage collection)
DBI::dbDisconnect(src$con)

}
\seealso{
\link{download_hydat}
}
