% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_stn_.R
\name{hy_stn_datum_unrelated}
\alias{hy_stn_datum_unrelated}
\title{hy_stn_datum_unrelated function}
\usage{
hy_stn_datum_unrelated(station_number = NULL,
  hydat_path = paste0(rappdirs::user_data_dir(), "\\\\Hydat.sqlite3"),
  prov_terr_state_loc = NULL)
}
\arguments{
\item{station_number}{Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory.
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}

\item{prov_terr_state_loc}{Province, state or territory. If this argument is omitted, the value of \code{station_number}
is returned. See \code{unique(allstations$prov_terr_state_loc)}}
}
\value{
A tibble of hy_stn_datum_unrelated
}
\description{
hy_stn_datum_unrelated look-up Table
}
\examples{
\donttest{
hy_stn_datum_unrelated()
}

}
