% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy.R
\name{hy_stn_datum_unrelated}
\alias{hy_stn_datum_unrelated}
\title{Extract station datum unrelated from HYDAT database}
\format{
A tibble with 4 variables:
\describe{
\item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
\item{DATUM_ID}{Unique code identifying a datum}
\item{Year_from}{First year of use}
\item{Year_to}{Last year of use}
}
}
\usage{
hy_stn_datum_unrelated(
  station_number = NULL,
  hydat_path = NULL,
  prov_terr_state_loc = NULL
)
}
\arguments{
\item{station_number}{A seven digit Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{hydat_path}{The path to the hydat database or NULL to use the default location
used by \link{download_hydat}. It is also possible to pass in an existing
\link[dplyr]{src_sqlite} such that the database only needs to be opened once per
user-level call.}

\item{prov_terr_state_loc}{Province, state or territory. If this argument is omitted, the value of \code{station_number}
is returned. See \code{unique(allstations$prov_terr_state_loc)}. Will also accept \code{CA} to return only Canadian stations.}
}
\value{
A tibble of hy_stn_datum_unrelated
}
\description{
hy_stn_datum_unrelated look-up Table
}
\examples{
\dontrun{
hy_stn_datum_unrelated()
}

}
