% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_id}
\alias{param_id}
\title{Parameter ID}
\format{
A tibble with 8 rows and 7 variables:
\itemize{
\item \strong{Parameter}: Numeric parameter code
\item \strong{Code}: Letter parameter code
\item \strong{Name_En}: Code name in English
\item \strong{Name_En}: Code name in French
\item \strong{Unit}: Parameter units
}
}
\usage{
param_id
}
\description{
A tibble of parameter id codes and their corresponding explanation/description specific to the ECCC webservice
}
\keyword{datasets}
