% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_plot.R
\name{plot}
\alias{plot}
\alias{plot.hy}
\title{plot hy object}
\usage{
\method{plot}{hy}(x = NULL, ...)
}
\arguments{
\item{x}{Object created by a daily data retrieval function}

\item{...}{passed to \code{plot}}
}
\description{
This method plots daily time series data from HYDAT. Functionality is very basic
and is intended for exploratory purposes only.
}
\examples{
\dontrun{
# One station
fraser <- hy_daily_flows("08MF005")
plot(fraser)
}

}
