% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_plot.R, R/realtime_plot.R
\name{plot}
\alias{plot}
\alias{plot.hy}
\alias{plot.realtime}
\title{plot hy object}
\usage{
\method{plot}{hy}(x = NULL, ...)

\method{plot}{realtime}(x = NULL, Parameter = c("Flow", "Level"), ...)
}
\arguments{
\item{x}{Object created by a hy_daily data retrieval function}

\item{...}{passed to \code{plot}}

\item{Parameter}{Parameter of interest. Either "Flow" or "Level". Defaults to "Flow".}

\item{x}{Object created by a realtime data retrieval function}

\item{...}{passed to \code{plot}}
}
\description{
This method plots daily time series data from HYDAT. Functionality is very basic
and is intended for exploratory purposes only.

This method plots daily time series data from the ECCC datamart. Functionality is very basic
and is intended for exploratory purposes only.
}
\examples{
\dontrun{
# One station
fraser <- hy_daily_flows("08MF005")
plot(fraser)
}

\dontrun{
# One station
fraser <- realtime_dd("08MF005")
plot(fraser)
}

}
