% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{realtime_dd}
\alias{realtime_dd}
\title{Download a tibble of realtime discharge data from the MSC datamart}
\usage{
realtime_dd(STATION_NUMBER = NULL, PROV_TERR_STATE_LOC)
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number. If this argument is omitted from the function call, the value of \code{PROV_TERR_STATE_LOC}
is returned.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. If this argument is omitted from the function call, the value of \code{STATION_NUMBER}
is returned.}
}
\value{
A tibble of water flow and level values. Time is return as UTC for consistency.
}
\description{
Download realtime discharge data from the Meteorological Service of Canada (MSC) datamart. The function will prioritize
downloading data collected at the highest resolution. In instances where data is not available at high (hourly or higher) resolution
daily averages are used. Currently, if a station does not exist or is not found, no data is returned.
}
\examples{
## Download from multiple provinces
realtime_dd(STATION_NUMBER=c("01CD005","08MF005"))

# To download all stations in Prince Edward Island:
realtime_dd(PROV_TERR_STATE_LOC = "PE")

}
\seealso{
Other realtime functions: \code{\link{realtime_stations}},
  \code{\link{realtime_ws}}, \code{\link{token_ws}}
}
