% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{realtime_network_meta}
\alias{realtime_network_meta}
\title{download a tibble of active realtime stations}
\usage{
realtime_network_meta(PROV_TERR_STATE_LOC)
}
\arguments{
\item{PROV_TERR_STATE_LOC}{Province/State/Territory or Location. See examples for list of available options. Use "ALL" for all stations.}
}
\description{
Returns all stations in the Realtime Water Survey of Canada hydrometric network operated by Environment and Cliamte Change Canada
}
\examples{
## Available inputs for PROV_TERR_STATE_LOC argument:
unique(realtime_network_meta(PROV_TERR_STATE_LOC = "ALL")$PROV_TERR_STATE_LOC)

realtime_network_meta(PROV_TERR_STATE_LOC = "BC")
}
