% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_plot.R
\name{realtime_plot}
\alias{realtime_plot}
\title{Convenience function to plot realtime data}
\usage{
realtime_plot(station_number = NULL, Parameter = c("Flow", "Level"))
}
\arguments{
\item{station_number}{A seven digit Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{Parameter}{Parameter of interest. Either "Flow" or "Level".}
}
\value{
A plot of recent realtime values
}
\description{
This is a quick and easy way to visualize a single station using base R graphics. 
More complicated plotting needs should consider using \code{ggplot2}.
}
\examples{
\dontrun{
realtime_plot("08MF005")
}

}
