% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{realtime_stations}
\alias{realtime_stations}
\title{download a tibble of active realtime stations}
\usage{
realtime_stations(PROV_TERR_STATE_LOC = NULL)
}
\arguments{
\item{PROV_TERR_STATE_LOC}{Province/State/Territory or Location. See examples for list of available options. 
realtime_stations() for all stations.}
}
\description{
An up to date dataframe of all stations in the Realtime Water Survey of Canada 
  hydrometric network operated by Environment and Climate Change Canada
}
\examples{
## Available inputs for PROV_TERR_STATE_LOC argument:
unique(realtime_stations()$PROV_TERR_STATE_LOC)

realtime_stations(PROV_TERR_STATE_LOC = "BC")
realtime_stations(PROV_TERR_STATE_LOC = c("QC","PE"))
}
\seealso{
Other realtime functions: \code{\link{realtime_dd}},
  \code{\link{realtime_ws}}, \code{\link{token_ws}}
}
