% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{realtime_ws}
\alias{realtime_ws}
\title{Download realtime data from the ECCC web service}
\format{A tibble with 6 variables:
\describe{
  \item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
  \item{Date}{Observation date and time. Formatted as a POSIXct class as UTC for consistency.}
  \item{Name_En}{Code name in English}
  \item{Value}{Value of the measurement.}
  \item{Unit}{Value units}
  \item{Grade}{future use}
  \item{Symbol}{future use}
  \item{Approval}{future use}
  \item{Parameter}{Numeric parameter code}
  \item{Code}{Letter parameter code}
}}
\usage{
realtime_ws(station_number, parameters = c(46, 16, 52, 47, 8, 5, 41, 18),
  start_date = Sys.Date() - 30, end_date = Sys.Date(), token)
}
\arguments{
\item{station_number}{Water Survey of Canada station number.}

\item{parameters}{parameter ID. Can take multiple entries. Parameter is a numeric code. See \code{param_id} for options. Defaults to all parameters.}

\item{start_date}{Need to be in YYYY-MM-DD. Defaults to 30 days before current date.}

\item{end_date}{Need to be in YYYY-MM-DD. Defaults to current date.}

\item{token}{generate by \code{token_ws()}}
}
\description{
Function to actually retrieve data from ECCC web service. Before using this function,
a token from \code{token_ws()} is needed. The maximum number of days that can be 
queried depends on other parameters being requested. If one station is requested, 18 
months of data can be requested. If you continually receiving errors when invoking this
function, reduce the number of observations (via station_number, parameters or dates) being requested.
}
\examples{
\dontrun{
token_out <- token_ws()

ws_08 <- realtime_ws(station_number = c("08NL071","08NM174"),
                         parameters = c(47, 5),
                         token = token_out)

fivedays <- realtime_ws(station_number = c("08NL071","08NM174"),
                         parameters = c(47, 5),
                         end_date = Sys.Date(), #today
                         start_date = Sys.Date() - 5, #five days ago
                         token = token_out)
}
}
\seealso{
Other realtime functions: \code{\link{realtime_dd}},
  \code{\link{realtime_stations}}, \code{\link{token_ws}}
}
