% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{station_choice}
\alias{station_choice}
\title{Function to chose a station based on consistent arguments for hydat functions.}
\usage{
station_choice(hydat_con, station_number, prov_terr_state_loc)
}
\arguments{
\item{hydat_con}{A database connection}

\item{station_number}{A seven digit Water Survey of Canada station number. If this argument is omitted, the value of \code{prov_terr_state_loc}
is returned.}

\item{prov_terr_state_loc}{Province, state or territory. If this argument is omitted, the value of \code{station_number}
is returned. See \code{unique(allstations$prov_terr_state_loc)}. Will also accept \code{CA} to return only Canadian stations.}
}
\description{
A function to avoid duplication in HYDAT functions.  This function is not intended for external use.
}
\keyword{internal}
