% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_timezones}
\alias{station_timezones}
\title{Time zones and offsets of each station}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7796 rows and 3 columns.}
\usage{
station_timezones
}
\description{
Time zones and offsets of each station calculated using the lutz package
}
\keyword{datasets}
