% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATIONS.R
\name{STATIONS}
\alias{STATIONS}
\title{Extract station information from the HYDAT database}
\usage{
STATIONS(hydat_path, STATION_NUMBER, PROV_TERR_STATE_LOC)
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory}

\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. See also for argument options.}
}
\value{
A tibble of stations and associated metadata
}
\description{
Provides wrapper to turn the STATIONS table in HYDAT into a tidy data frame. \code{STATION_NUMBER} and 
\code{PROV_TERR_STATE_LOC} must both be supplied. When STATION_NUMBER="ALL" the PROV_TERR_STATE_LOC argument decides 
where those stations come from.
}
\examples{
## Two stations
STATIONS(STATION_NUMBER = c("08CG001", "08CE001"), 
   PROV_TERR_STATE_LOC = "BC", hydat_path = "H:/Hydat.sqlite3")
## ALL stations from PEI
STATIONS(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")

}
\seealso{
\code{PROV_TERR_STATE_LOC} abbreviations
\itemize{
\item "QC" 
\item "ME" 
\item "NB" 
\item "PE" 
\item "NS" 
\item "MN" 
\item "ON" 
\item "MI" 
\item "NL" 
\item "MB" 
\item "AB" 
\item "MT" 
\item "SK" 
\item "ND" 
\item "NU" 
\item "NT" 
\item "BC" 
\item "YT" 
\item "AK" 
\item "WA" 
\item "ID"
}
}
