% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATIONS.R
\name{STATIONS}
\alias{STATIONS}
\title{Station Hydat wrapper}
\usage{
STATIONS(hydat_path = "H:/Hydat.sqlite3",
  STATION_NUMBER_SEL = STATION_NUMBER)
}
\arguments{
\item{hydat_path}{Directory to the hydat database}

\item{STATION_NUMBER_SEL}{Water Survey of Canada station number}
}
\description{
Provides wrapper to turn the STATIONS table into a tidy data frame
}
\examples{

## One station
STATIONS(STATION_NUMBER = "08CG001")
## Two stations
STATIONS(STATION_NUMBER = c("08CG001", "08CE001"))
}
