% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STN_REGULATION.R
\name{STN_REGULATION}
\alias{STN_REGULATION}
\title{Extract station regulation from the HYDAT database}
\usage{
STN_REGULATION(hydat_path, STATION_NUMBER = "ALL")
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory}

\item{STATION_NUMBER}{Water Survey of Canada station number. No default. Can also take the "ALL" argument.}
}
\value{
A tibble of stations, years of regulation and the regulation status
}
\description{
Provides wrapper to turn the STN_REGULATION table in HYDAT into a tidy data frame. \code{STATION_NUMBER} and 
\code{PROV_TERR_STATE_LOC} must both be supplied. When STATION_NUMBER="ALL" the PROV_TERR_STATE_LOC argument decides 
where those stations come from.
}
\examples{
\donttest{
## Two stations
STN_REGULATION(STATION_NUMBER = c("08CG001", "08CE001"), hydat_path = "H:/Hydat.sqlite3")

## ALL stations from PEI
STN_REGULATION(STATION_NUMBER = "ALL", hydat_path = "H:/Hydat.sqlite3")
}

}
