% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{tidyeval}
\alias{tidyeval}
\alias{quo}
\alias{quos}
\alias{enquo}
\alias{sym}
\alias{syms}
\alias{ensym}
\alias{expr}
\alias{exprs}
\alias{enexpr}
\alias{quo_name}
\title{Tidy eval helpers}
\description{
These functions provide tidy eval-compatible ways to capture
symbols (`sym()`, `syms()`, `ensym()`), expressions (`expr()`,
`exprs()`, `enexpr()`), and quosures (`quo()`, `quos()`, `enquo()`).
To learn more about tidy eval and how to use these tools, read
<http://rlang.tidyverse.org/articles/tidy-evaluation.html>
}
\keyword{internal}
