% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhydat.R
\docType{package}
\name{tidyhydat}
\alias{tidyhydat}
\alias{tidyhydat-package}
\title{Make HYDAT data tidy}
\description{
This package provides several quick and easy functions to extract data from the HYDAT database and turn that into tidy data. 
Most functions reflects a table in the HYDAT database. Realtime data acquisition is also provided.

\code{tidyhydat} package
}
\references{
For more information on tidy data please see
\itemize{
 \item Wickham, Hadley. 2014. Tidy Data. The Journal of Statistical Software. 59. \url{https://www.jstatsoft.org/article/view/v059i10}
 \item tidy data vignette: \url{https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html}
 }
 
 To download the latest version of hydat please follow this link: 
 \itemize{
  \item \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/Hydat_sqlite3_20170418.zip}
  }
}
