% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhydat.R
\docType{package}
\name{tidyhydat}
\alias{tidyhydat}
\alias{tidyhydat-package}
\title{Extract tidy river data}
\description{
tidyhydat provides functions to extract river data from Water Survey of Canada sources and make it tidy.
}
\references{
For more information on tidy data please see
\itemize{
 \item Wickham, Hadley. 2014. Tidy Data. The Journal of Statistical Software. 59. \url{https://www.jstatsoft.org/article/view/v059i10}
 \item tidy data vignette: \url{https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html}
 }

 To download the latest version of hydat please follow this link:
 \itemize{
  \item \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/}
  }
}
