% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VERSION}
\alias{VERSION}
\title{Version number of HYDAT}
\usage{
VERSION(hydat_path = NULL)
}
\arguments{
\item{hydat_path}{Directory to the hydat database. Can be set as "Hydat.sqlite3" which will look for Hydat in the working directory. 
The hydat path can also be set in the \code{.Renviron} file so that it doesn't have to specified every function call. The path should 
set as the variable \code{hydat}. Open the \code{.Renviron} file using this command: \code{file.edit("~/.Renviron")}.}
}
\value{
version number
}
\description{
A function to get version number of hydat
}
