% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical-webservice.R
\name{ws_daily_flows}
\alias{ws_daily_flows}
\alias{ws_daily_levels}
\title{Download historical data from the ECCC web service}
\format{
A tibble with 6 variables:
\describe{
\item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
\item{Date}{Observation date and time. Formatted as a POSIXct class as UTC for consistency.}
\item{Parameter}{Type of parameter}
\item{Value}{Value of the measurement.}
\item{Symbol}{future use}
}
}
\usage{
ws_daily_flows(
  station_number,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date()
)

ws_daily_levels(
  station_number,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date()
)
}
\arguments{
\item{station_number}{Water Survey of Canada station number.}

\item{start_date}{Accepts YYYY-MM-DD.
If only start date is supplied (i.e. YYYY-MM-DD) values are returned from the start of that day.
Defaults to 365 days before current date.}

\item{end_date}{Accepts either YYYY-MM-DD.
If only a date is supplied (i.e. YYYY-MM-DD) values are returned from the end of that day.
Defaults to current date.}
}
\description{
Function to retrieve historical data from ECCC web service. This data is
the same as HYDAT data but provides the convenience of not having to download
the HYDAT database. This function when a smaller amount of data is needed. If
you need lots of data, consider the \code{hy_} family of functions
}
\examples{
\dontrun{

flow_data <- ws_daily_flows(
  station_number = c("08NL071", "08NM174")
)

level_data <- ws_daily_level(
  station_number = c("08NL071", "08NM174")
)
}
}
\seealso{
hy_daily_flows
}
